Moved CloudfixIntegration to a new package
