+++ .gitignore
@@ -5,3 +5,4 @@ recos.txt
 coverage.out
 mappingFile.json
 tagsID.txt
+.vscode
